Modify imu function definition

+++ m5atom-matrix/include/imu.h
@@ -2,7 +2,7 @@
 #define __IMU__
 
 void init_imu();
-void imu_getaccel(uint16_t *accx, uint16_t *accy, uint16_t *accz );
+void imu_getaccel(float *accx, float *accy, float *accz );
 void imu_calibrate();
 
 #endif
