+++ m5atom-matrix/include/scratchlink.h
@@ -5,7 +5,7 @@ const int ScratchLink_Say = 0x81;
 const int ScratchLink_Led = 0x82;
 
 extern void init_scratchlink(const char *devicename);
-extern void scratchlink_update(uint16_t accx, uint16_t accy, bool buttonA, bool buttonB, bool touch0, bool touch1, bool touch2, int gesture);
+extern void scratchlink_update(float accx, float accy, bool buttonA, bool buttonB, bool touch0, bool touch1, bool touch2, int gesture);
 extern void scratchlink_callback(byte command, byte *payload, unsigned int size);
 extern bool scratchlink_isconnecting();
